requirements.txt
fastapi==0.115.4
httpx==0.28.1
pydantic==2.9.2
pytest==8.3.3
SQLAlchemy==2.0.36
starlette==0.41.2
uvicorn==0.32.0
psycopg2==2.9.10
python-jose==3.3.0
bcrypt==4.2.0
passlib==1.7.4
dnspython==2.7.0
email-validator==2.2.0
alembic==1.13.3
pydantic-settings==2.6.1

python-jose==3.3.0
bcrypt==4.2.0
passlib==1.7.4

psycopg2-binary==2.9.10
asyncpg==0.30.0
phonenumbers==8.13.51